Move DirtyQubitTag to new tags package
