+++ projectq/cengines/__init__.py
@@ -14,7 +14,6 @@ from ._basics import (BasicEngine,
                       LastEngineException,
                       ForwarderEngine)
 from ._cmdmodifier import CommandModifier
-from ._dirtyqubit import DirtyQubitTag
 from ._ibmcnotmapper import IBMCNOTMapper
 from ._main import (MainEngine,
                     NotYetMeasuredError,
