+++ projectq/cengines/_basics.py
@@ -13,6 +13,7 @@
 import projectq.cengines
 from projectq.ops import Allocate, Deallocate
 from projectq.ops import Command
+from projectq.tags import DirtyQubitTag
 from projectq.types import Qubit, Qureg
 
 
@@ -108,12 +109,12 @@ class BasicEngine(object):
         new_id = self.main_engine.get_new_qubit_id()
         qb = Qureg([Qubit(self, new_id)])
         if dirty:
-            if self.is_meta_tag_supported(projectq.cengines.DirtyQubitTag):
+            if self.is_meta_tag_supported(DirtyQubitTag):
                 oldnext = self.next_engine
 
                 def cmd_modifier(cmd):
                     assert(cmd.gate == Allocate)
-                    cmd.tags += [projectq.cengines.DirtyQubitTag()]
+                    cmd.tags += [DirtyQubitTag()]
                     return cmd
                 self.next_engine = projectq.cengines.CommandModifier(
                     cmd_modifier)
@@ -161,7 +162,7 @@ class BasicEngine(object):
 
             def cmd_modifier(cmd):
                 assert(cmd.gate == Deallocate)
-                cmd.tags += [projectq.cengines.DirtyQubitTag()]
+                cmd.tags += [DirtyQubitTag()]
                 return cmd
             self.next_engine = projectq.cengines.CommandModifier(
                 cmd_modifier)
