+++ projectq/meta/_dagger_test.py
@@ -12,15 +12,15 @@
 
 """Tests for projectq.meta._dagger.py"""
 
-import pytest
 import types
+
+import pytest
 
 from projectq import MainEngine
 from projectq.cengines import DummyEngine
-from projectq.ops import CNOT, X, Rx, H, Allocate, Deallocate
-from projectq.meta import DirtyQubitTag
-
 from projectq.meta import _dagger
+from projectq.ops import CNOT, X, Rx, H, Allocate, Deallocate
+from projectq.tags import DirtyQubitTag
 
 
 def test_dagger_with_dirty_qubits():
