+++ projectq/meta/_loop_test.py
@@ -12,16 +12,16 @@
 
 """Tests for projectq.meta._loop.py"""
 
-import pytest
 import types
-
 from copy import deepcopy
-from projectq import MainEngine
-from projectq.meta import ComputeTag, DirtyQubitTag
-from projectq.cengines import DummyEngine
-from projectq.ops import H, CNOT, X, FlushGate, Allocate, Deallocate
 
+import pytest
+
+from projectq import MainEngine
+from projectq.cengines import DummyEngine
+from projectq.meta import ComputeTag
 from projectq.meta import _loop
+from projectq.ops import H, CNOT, X, FlushGate, Allocate, Deallocate
 
 
 def test_loop_tag():
