+++ projectq/tags/__init__.py
@@ -1,0 +1,1 @@
+from ._dirtyqubit import DirtyQubitTag
